@@ -1,13 +1,13 @@
 module github.com/wednesday-solutions/picky
 
-go 1.18
+go 1.20
 
 require (
 	github.com/aymerick/raymond v2.0.2+incompatible
 	github.com/enescakir/emoji v1.0.0
 	github.com/fatih/color v1.15.0
-	github.com/iancoleman/strcase v0.2.0
-	github.com/rivo/tview v0.0.0-20230406072732-e22ce9588bb4
+	github.com/iancoleman/strcase v0.3.0
+	github.com/rivo/tview v0.0.0-20230814110005-ccc2c8119703
 	github.com/schollz/progressbar/v3 v3.13.1
 	github.com/spaceweasel/promptui v0.8.1
 	github.com/spf13/cobra v1.7.0
@@ -22,14 +22,14 @@ require (
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/lunixbochs/vtclean v1.0.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.18 // indirect
-	github.com/mattn/go-runewidth v0.0.14 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
 	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	github.com/stretchr/testify v1.8.2 // indirect
-	golang.org/x/sys v0.7.0 // indirect
-	golang.org/x/term v0.7.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
+	github.com/stretchr/testify v1.8.4 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/term v0.11.0 // indirect
+	golang.org/x/text v0.12.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
